Move DownloadMaps to FGEImage
